+++ src/main/java/io/supertokens/inmemorydb/Start.java
@@ -497,6 +497,23 @@ public class Start implements SessionSQLStorage, EmailPasswordSQLStorage, EmailV
     }
 
     @Override
+    public void updateUsersEmail_Transaction(TransactionConnection conn, String userId, String email)
+            throws StorageQueryException, DuplicateEmailException {
+        Connection sqlConn = (Connection) conn.getConnection();
+
+        try {
+            EmailPasswordQueries.updateUsersEmail_Transaction(this, sqlConn, userId, email);
+        } catch (SQLException e) {
+            if (e.getMessage()
+                    .equals("[SQLITE_CONSTRAINT]  Abort due to constraint violation (UNIQUE constraint failed: "
+                            + Config.getConfig(this).getEmailPasswordUsersTable() + ".email)")) {
+                throw new DuplicateEmailException();
+            }
+            throw new StorageQueryException(e);
+        }
+    }
+
+    @Override
     public UserInfo getUserInfoUsingId_Transaction(TransactionConnection con, String userId)
             throws StorageQueryException {
         Connection sqlCon = (Connection) con.getConnection();